requirements.txt
# 🛡️ DETERMINISTIC CLINICAL BASELINE (Verified Versions)

# Core Deep Learning (Pinned to latest stable medical verified)
torch==2.10.0
torchvision==0.25.0
ultralytics==8.4.31

# Medical Computer Vision & Array Manipulation
opencv-python-headless==4.10.0.84
numpy==1.26.4
openslide-python==1.3.1

# High-Performance API Backend
fastapi==0.128.0
uvicorn[standard]==0.32.1
pydantic==2.12.5

# Configuration & Security utilities
pyyaml==6.0.1
python-multipart==0.0.12

# Clinical Reporting
fpdf2==2.8.1
